External: merge json source files
